git-svn-id: file:///home/gumz3/Documents/Braker/svn/braker/trunks@536 52e7708f-445d-428a-8816-bceb3badd3a6

+++ scripts/braker.pl
@@ -3094,12 +3094,10 @@ sub training {
             . (localtime)
             . ": Deleting intermediate training gene structure files:\n"
             . "rm $trainGb2 $trainGb3 $otherfilesDir/traingenes.good.nr.fa $otherfilesDir/nonred.loci.lst $otherfilesDir/traingenes.good.gtf $otherfilesDir/etrain.bad.lst $goodLstFile\n";
-        my $ccc=0;
         foreach ( ($trainGb2, $trainGb3, "$otherfilesDir/traingenes.good.nr.fa", "$otherfilesDir/nonred.loci.lst", "$otherfilesDir/traingenes.good.gtf", "$otherfilesDir/etrain.bad.lst", $goodLstFile) ) {
-           print "Will try to delete file: $ccc\n";
-           print $_."\n";
-          unlink ( $_ ) or die ("ERROR in file " . __FILE__ ." at line ". __LINE__ ."\nFailed to delete file $_!\n");
-          $ccc++;
+            if (-e $_ ) {
+                unlink ( $_ ) or die ("ERROR in file " . __FILE__ ." at line ". __LINE__ ."\nFailed to delete file $_!\n");
+            }
         }
 
         # split into training and test set
@@ -3369,7 +3367,7 @@ sub training {
                 }
                 if ($nice) {
                     $perlCmdString
-                        .= "perl $string --nice --rounds=$rounds --species=$species --AUGUSTUS_CONFIG_PATH=$AUGUSTUS_CONFIG_PATH --onlytrain=$otherfilesDir/train.gb.train.train --cpus=$CPU $otherfilesDir/train.gb.train.test 1>$stdoutfile 2>$errorfile";
+                        .= "perl $string --nice --rounds=$rounds --species=$species --AUGUSTUS_CONFIG_PATH=$AUGUSTUS_CONFIG_PATH --onlytrain=$otherfilesDir/train.gb.train.train --cleanup=1 --cpus=$CPU $otherfilesDir/train.gb.train.test 1>$stdoutfile 2>$errorfile";
                 }
                 else {
                     $perlCmdString
@@ -3948,6 +3946,9 @@ sub clean_up {
             print LOG "rm $otherfilesDir/$file\n";
             unlink( "$otherfilesDir/$file" );
         }
+    }
+    if(-e "$otherfilesDir/seqlist"){
+        unlink ( "$otherfilesDir/seqlist" );
     }
 }
